requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
backcall=0.2.0=pypi_0
blas=1.0=mkl
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.9.24=py37h06a4308_0
cffi=1.15.0=py37hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=37.0.1=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pypi_0
expat=2.4.4=h295c915_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
glog=0.3.1=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipython=7.34.0=pypi_0
jedi=0.18.1=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=py37h295c915_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libffi=3.3=he6710b0_2
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=12.1.0=h8d9b700_16
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.5.1=py37h06a4308_1
matplotlib-base=3.5.1=py37ha18d171_1
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pyhd3eb1b0_0
numpy=1.21.5=py37he7a7128_2
numpy-base=1.21.5=py37hf524024_2
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py37h22f2fdc_0
pip=21.2.2=py37h06a4308_0
prompt-toolkit=3.0.31=pypi_0
protobuf=3.20.1=pypi_0
ptyprocess=0.7.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.13.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py37h06a4308_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_1
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-gflags=3.1.2=pypi_0
pytorch=1.11.0=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2022.4=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=py37hc147768_0
setuptools=61.2.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
sqlite=3.38.3=hc218d9a_0
tensorboardx=2.5.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_1
torchaudio=0.11.0=py37_cu113
torchvision=0.12.0=py37_cu113
tornado=6.1=py37h27cfd23_0
tqdm=4.64.0=py37h06a4308_0
traitlets=5.4.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=py37h06a4308_0
wcwidth=0.2.5=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
